v3.0.1 [NEW] lib.rs - Creates a library of hashing functions which makes data interpolation easier. [FIX] fixed issue with csv output not being all screwed up.
Overall POC is more of the concept with this commit. As there are many things that will need to be rewritten as there is no database functions that are currently implemented. This will be another function that will be considered as the project goes on.
A current issue that will be addressed is taking output from the last input from lib.rs and taking that and putting it into the csv file. This problem is currently being investigated.